update to Minecraft 1.8.9

+++ src/main/java/com/github/rainang/logicgates/block/BlockDiode2In.java
@@ -41,7 +41,7 @@ public abstract class BlockDiode2In extends BlockDiode {
 
 	@Override
 	public int getInputState(IBlockState state) {
-		return (Integer)state.getValue(INPUT);
+		return state.getValue(INPUT);
 	}
 
 	@Override
